% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositiADA-package.R
\docType{package}
\name{compositiADA-package}
\alias{compositiADA}
\alias{compositiADA-package}
\title{compositiADA: I dati per gli indicatori compositi}
\description{
Il pacchetto contiene i dati per i compositi 2019.
}
\author{
\strong{Maintainer}: Raffaele Morelli \email{raffaele.morelli@isprambiente.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
