% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obiettivo-01a.R
\docType{data}
\name{Obiettivo-01a}
\alias{Obiettivo-01a}
\title{Data of nominal brown shrimp prices by size}
\format{
A tibble with 243 rows and 10 variables:
\describe{
\item{Year}{dbl Year price was recorded}
\item{Month}{dbl Month price was recorded. Ranges from 1-12 for January - December}
\item{Pieces}{dbl denoting monthly price of pieces of shrimp totalling a pound}
}
}
\usage{
Obiettivo-01a
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\keyword{datasets}
